push to cloud

+++ Assets/default.vert
@@ -1,10 +1,10 @@
 #version 450 core
 layout (location = 0) in vec3 aPos;
 layout (location = 1) in mat4 aXform;
-layout (location = 5) in uint aIsCollide;
+layout (location = 5) in float aIsCollide;
 
 uniform mat4 uProjViewMatrix;
-flat out uint oIsCollide;
+flat out float oIsCollide;
 
 void main()
 {
